# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  /** Number of active AI store managers */
  activeStores: number;
  /** Total orders/leads captured across all stores */
  totalLeads: number;
  /** Orders received today */
  messagesToday: number;
  /** Total stores created */
  totalStores: number;
}

export interface StoreWithStats {
  id: string;
  storeName: string;
  botUsername: string;
  isActive: boolean;
  source?: string;
  orderCount: number;
  createdAt: string;
}

export interface OrderSummary {
  id: string;
  storeName: string;
  customerName: string;
  customerPhone: string;
  customerAddress?: string;
  totalPrice: string;
  status: string;
  source: string;
  createdAt: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  messages: ChatMessage[];
  model?: string;
}

export interface ChatResponse {
  content: string;
  model: string;
}