# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI chat tester (non-streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
components:
  schemas:
...
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
      required:
        - role
        - content
    ChatInput:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        model:
          type: string
          default: "Qwen/Qwen2.5-72B-Instruct"
      required:
        - messages
    ChatResponse:
      type: object
      properties:
        content:
          type: string
        model:
          type: string
      required:
        - content
        - model